Using the correct namespace on the User model after moving it. https://github.com/julisana/laravel-recipes/issues/10
 Changes to be committed:
	modified:   app/Http/routes.php
	modified:   app/Models/User.php
